frontend/requirements.txt
streamlit
requests
Pillow
numpy
easyocr
beautifulsoup4